Cleanup: generalize all HTTP commands into one module
This is the first commit which just copies get and post commands into an http/http.go module verbatim.
Future commits will generalize on it.
